make card sane
