+++ src/Components/Card/NotRecruiting.js
@@ -1,0 +1,23 @@
+import React from "react";
+const NotRecruiting = () => {
+  return (
+    <>
+      <div className="closed">
+        <svg
+          aria-labelledby="closed"
+          xmlns="http://www.w3.org/2000/svg"
+          width="48"
+          height="48"
+          viewBox="0 0 48 48"
+        >
+          <title id="closed" lang="en">
+            Not Recruiting
+          </title>
+          <path d="M38 12.83L35.17 10 24 21.17 12.83 10 10 12.83 21.17 24 10 35.17 12.83 38 24 26.83 35.17 38 38 35.17 26.83 24z" />
+        </svg>
+      </div>
+      <div>Not Recruiting</div>
+    </>
+  );
+};
+export default NotRecruiting;
