+++ src/Components/Card/Recruiting.js
@@ -1,0 +1,23 @@
+import React from "react";
+const Recruiting = () => {
+  return (
+    <>
+      <div className="open">
+        <svg
+          aria-labelledby="open"
+          xmlns="http://www.w3.org/2000/svg"
+          width="48"
+          height="48"
+          viewBox="0 0 48 48"
+        >
+          <title id="open" lang="en">
+            Recruiting
+          </title>
+          <path d="M18 32.34L9.66 24l-2.83 2.83L18 38l24-24-2.83-2.83z" />
+        </svg>
+      </div>
+      <div>Recruiting</div>
+    </>
+  );
+};
+export default Recruiting;
